pyproject.toml
[tool.poetry]
name = "alexofechoes_taskmanager"
version = "0.1.0"
description = "Taskmanager - hexlet fourth project"
authors = ["Alexander Bukharin <alexawm182@gmail.com>"]
repository = "https://github.com/alexofechoes/python-project-lvl4"

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.6"
django-bootstrap4 = "^1.1.1"
django-bootstrap-breadcrumbs = "^0.9.2"
django-crispy-forms = "^1.9.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.5"
python-dotenv = "^0.13.0"
rollbar = "^0.15.0"
whitenoise = "^5.0.1"
django-filter = "^2.3.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
django-stubs = "^1.5.0"
flake8 = "^3.7.9"
isort = "^4.3.21"
mypy = "^0.770"
pytest = "^5.2"
pytest-cov = "^2.8.1"
wemake-python-styleguide = "^0.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
